+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - When showing the correct answer, replace inserted or removed spaces with an underscore so it is more clear what the user would have needed to type differently to enter the correct answer. Fixes [#96](https://github.com/fniessink/toisto/issues/96).
 - Apostrophes were ignored when checking answers. Fixes [#97](https://github.com/fniessink/toisto/issues/97).
+- Toisto would show empty strings as meaning if the quizzed concept has no label in the user's language. Fixes [#101](https://github.com/fniessink/toisto/issues/101).
 
 ## v0.3.0 - 2022-12-04